+++ package.json
@@ -13,6 +13,7 @@
     "cli1_powershell": "$env:RSS_value1='value1'; $env:RSS_value2='value2'; node ./src/cli/env.js",
     "cli1_bash": "RSS_name1=value1 RSS_name2=value2 node ./src/cli/env.js",
     "cli2": "node ./src/cli/args.js --propName value --prop2Name value2",
+    "module": "node ./src/modules/cjsToEsm.mjs",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
